BeaconOracle.sol add pendingBalances  and variables use uint256

+++ src/oracles/ReportUtils.sol
@@ -8,8 +8,8 @@ library ReportUtils {
      */
     function compressReportData(
         bytes32 _validatorRankingRoot,
-        uint128 _beaconBalance,
-        uint32 _beaconValidators,
+        uint256 _beaconBalance,
+        uint256 _beaconValidators,
         uint16 _sameCount
     ) internal pure returns (bytes memory) {
         return abi.encode(_validatorRankingRoot, _beaconBalance, _beaconValidators, _sameCount);
@@ -20,8 +20,8 @@ library ReportUtils {
      * @param {bytes memory} data compressed data
      * @return decompressed data
      */
-    function decompressReportData(bytes memory data) internal pure returns (bytes32, uint128, uint32, uint16) {
-        return abi.decode(data, (bytes32, uint128, uint32, uint16));
+    function decompressReportData(bytes memory data) internal pure returns (bytes32, uint256, uint256, uint16) {
+        return abi.decode(data, (bytes32, uint256, uint256, uint16));
     }
 
     /**
@@ -31,10 +31,10 @@ library ReportUtils {
     function isReportDifferentAndCount(
         bytes memory value,
         bytes32 _validatorRankingRoot,
-        uint128 _beaconBalance,
-        uint32 _beaconValidators
+        uint256 _beaconBalance,
+        uint256 _beaconValidators
     ) internal pure returns (bool, uint16) {
-        (bytes32 root, uint128 balance, uint32 validators, uint16 sameCount) = decompressReportData(value);
+        (bytes32 root, uint256 balance, uint256 validators, uint16 sameCount) = decompressReportData(value);
         bool isDifferent =
             !(root == _validatorRankingRoot && balance == _beaconBalance && validators == _beaconValidators);
         return (isDifferent, sameCount);
